# openapi-generated-aspnetcore/src/Org.OpenAPITools/Controllers/PostsApi.cs
/*
 * PIM-SWKS SoSe2024 REST API
 *
 * PIM-SWKS SoSe2024 REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class PostsApiController : ControllerBase
    { 
        /// <summary>
        /// Create a new post
        /// </summary>
        /// <param name="post"></param>
        /// <response code="201">Created</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        [HttpPost]
        [Route("/posts")]
        [Authorize(Policy = "apikeyAuth")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreatePost")]
        [SwaggerResponse(statusCode: 201, type: typeof(Post), description: "Created")]
        public virtual IActionResult CreatePost([FromBody]Post post)
        {

            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Post));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Post>(exampleJson)
            : default(Post);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete a post by ID
        /// </summary>
        /// <param name="id">ID of the post to delete</param>
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">If the post does not exist</response>
        [HttpDelete]
        [Route("/posts/{id}")]
        [Authorize(Policy = "apikeyAuth")]
        [ValidateModelState]
        [SwaggerOperation("DeletePostById")]
        [SwaggerResponse(statusCode: 200, type: typeof(Post), description: "OK")]
        public virtual IActionResult DeletePostById([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Post));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Post>(exampleJson)
            : default(Post);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns a post by ID
        /// </summary>
        /// <param name="id">ID of the post to return</param>
        /// <response code="200">OK</response>
        /// <response code="404">If the post does not exist</response>
        [HttpGet]
        [Route("/posts/{id}")]
        [ValidateModelState]
        [SwaggerOperation("GetPostById")]
        [SwaggerResponse(statusCode: 200, type: typeof(Post), description: "OK")]
        public virtual IActionResult GetPostById([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Post));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Post>(exampleJson)
            : default(Post);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns all posts
        /// </summary>
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/posts")]
        [ValidateModelState]
        [SwaggerOperation("GetPosts")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Post>), description: "OK")]
        public virtual IActionResult GetPosts()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Post>));
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<Post>>(exampleJson)
            : default(List<Post>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns all posts authored by this user
        /// </summary>
        /// <param name="id">ID of the user to return posts for</param>
        /// <response code="200">OK</response>
        /// <response code="404">If the user does not exist</response>
        [HttpGet]
        [Route("/users/{id}/posts")]
        [ValidateModelState]
        [SwaggerOperation("GetPostsForUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Post>), description: "OK")]
        public virtual IActionResult GetPostsForUser([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Post>));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<Post>>(exampleJson)
            : default(List<Post>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update a post by ID
        /// </summary>
        /// <param name="id">ID of the post to update</param>
        /// <param name="post"></param>
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">If the post does not exist</response>
        [HttpPut]
        [Route("/posts/{id}")]
        [Authorize(Policy = "apikeyAuth")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("UpdatePostById")]
        [SwaggerResponse(statusCode: 200, type: typeof(Post), description: "OK")]
        public virtual IActionResult UpdatePostById([FromRoute (Name = "id")][Required]long id, [FromBody]Post post)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Post));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Post>(exampleJson)
            : default(Post);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}