# openapi-generated-aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "PIM-SWKS SoSe2024 REST API",
    "title" : "PIM-SWKS SoSe2024 REST API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:{Port}",
    "variables" : {
      "Port" : {
        "default" : "18080",
        "description" : "The port number (18080 - rest-jakarta, 28080 - rest-springboot, 38080 - rest-aspnet)",
        "enum" : [ "18080", "28080", "38080", "8080" ]
      }
    }
  }, {
    "description" : "Production server",
    "url" : "https://pim-swks.nekanu.com"
  } ],
...
  "paths" : {
    "/posts" : {
      "get" : {
        "operationId" : "getPosts",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Post"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns all posts",
        "tags" : [ "posts" ]
      },
      "post" : {
        "operationId" : "createPost",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Post"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Post"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Create a new post",
        "tags" : [ "posts" ]
      }
    },
    "/posts/{id}" : {
      "delete" : {
        "operationId" : "deletePostById",
        "parameters" : [ {
          "description" : "ID of the post to delete",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Post"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "If the post does not exist"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Delete a post by ID",
        "tags" : [ "posts" ]
      },
      "get" : {
        "operationId" : "getPostById",
        "parameters" : [ {
          "description" : "ID of the post to return",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Post"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "description" : "If the post does not exist"
          }
        },
        "summary" : "Returns a post by ID",
        "tags" : [ "posts" ]
      },
      "put" : {
        "operationId" : "updatePostById",
        "parameters" : [ {
          "description" : "ID of the post to update",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Post"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Post"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "If the post does not exist"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Update a post by ID",
        "tags" : [ "posts" ]
      }
    },
...
    "/users/{id}/posts" : {
      "get" : {
        "operationId" : "getPostsForUser",
        "parameters" : [ {
          "description" : "ID of the user to return posts for",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Post"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "summary" : "Returns all posts authored by this user",
        "tags" : [ "posts", "users" ]
      }
    }
  },
  "components" : {
...
    "schemas" : {
      "Post" : {
        "example" : {
          "id" : 1,
          "title" : "Hello World",
          "authorId" : 1,
          "content" : "This is a post"
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "format" : "uint64",
            "readOnly" : true,
            "type" : "integer"
          },
          "title" : {
            "example" : "Hello World",
            "maxLength" : 512,
            "minLength" : 3,
            "type" : "string"
          },
          "content" : {
            "example" : "This is a post",
            "format" : "markdown",
            "type" : "string"
          },
          "authorId" : {
            "description" : "ID of the author of the post",
            "example" : 1,
            "format" : "uint64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apikeyAuth" : {
        "in" : "header",
        "name" : "X-API-KEY",
        "type" : "apiKey"
      }
    }
  }
}